requirements.txt
google-cloud-texttospeech
google-cloud-speech
google.cloud
pyaudio
simpleaudio
vertexai
google-cloud-aiplatform
googlemaps
pygame
ultralytics
ollama
py3-tts
pyttsx3
